backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aiodns==4.0.0
    # via -r requirements.in
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.3
    # via -r requirements.in
aiosignal==1.4.0
    # via aiohttp
alembic==1.18.1
    # via -r requirements.in
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   httpx
    #   starlette
attrs==25.4.0
    # via aiohttp
certifi==2026.1.4
    # via
    #   httpcore
    #   httpx
cffi==2.0.0
    # via pycares
choreographer==1.2.1
    # via kaleido
click==8.3.1
    # via
    #   rq
    #   typer
    #   typer-slim
    #   uvicorn
coverage[toml]==7.13.2
    # via pytest-cov
croniter==6.0.0
    # via rq
cuda-bindings==12.9.4
    # via torch
cuda-pathfinder==1.3.3
    # via cuda-bindings
dateparser==1.2.2
    # via -r requirements.in
fastapi==0.128.0
    # via
    #   -r requirements.in
    #   fastapi-cache2
fastapi-cache2[redis]==0.2.2
    # via -r requirements.in
feedgen==1.0.0
    # via -r requirements.in
feedparser==6.0.12
    # via -r requirements.in
filelock==3.20.3
    # via
    #   huggingface-hub
    #   torch
    #   transformers
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2026.1.0
    # via
    #   huggingface-hub
    #   torch
greenlet==3.3.1
    # via sqlalchemy
gunicorn==24.1.1
    # via -r requirements.in
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
hf-xet==1.2.0
    # via huggingface-hub
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   -r requirements.in
    #   huggingface-hub
huggingface-hub==1.3.4
    # via
    #   tokenizers
    #   transformers
idna==3.11
    # via
    #   anyio
    #   httpx
    #   yarl
iniconfig==2.3.0
    # via pytest
jinja2==3.1.6
    # via torch
joblib==1.5.3
    # via scikit-learn
kaleido==1.2.0
    # via -r requirements.in
logistro==2.0.1
    # via
    #   choreographer
    #   kaleido
loguru==0.7.3
    # via -r requirements.in
lxml==6.0.2
    # via
    #   -r requirements.in
    #   feedgen
mako==1.3.10
    # via alembic
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via
    #   jinja2
    #   mako
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
multidict==6.7.1
    # via
    #   aiohttp
    #   yarl
narwhals==2.15.0
    # via plotly
networkx==3.6.1
    # via torch
numpy==2.4.1
    # via
    #   -r requirements.in
    #   pandas
    #   scikit-learn
    #   scipy
    #   transformers
nvidia-cublas-cu12==12.8.4.1
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.8.90
    # via torch
nvidia-cuda-nvrtc-cu12==12.8.93
    # via torch
nvidia-cuda-runtime-cu12==12.8.90
    # via torch
nvidia-cudnn-cu12==9.10.2.21
    # via torch
nvidia-cufft-cu12==11.3.3.83
    # via torch
nvidia-cufile-cu12==1.13.1.3
    # via torch
nvidia-curand-cu12==10.3.9.90
    # via torch
nvidia-cusolver-cu12==11.7.3.90
    # via torch
nvidia-cusparse-cu12==12.5.8.93
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cusparselt-cu12==0.7.1
    # via torch
nvidia-nccl-cu12==2.27.5
    # via torch
nvidia-nvjitlink-cu12==12.8.93
    # via
    #   nvidia-cufft-cu12
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
    #   torch
nvidia-nvshmem-cu12==3.4.5
    # via torch
nvidia-nvtx-cu12==12.8.90
    # via torch
orjson==3.11.5
    # via kaleido
packaging==26.0
    # via
    #   gunicorn
    #   huggingface-hub
    #   kaleido
    #   plotly
    #   pytest
    #   transformers
pandas==3.0.0
    # via -r requirements.in
pendulum==3.1.0
    # via fastapi-cache2
plotly==6.5.2
    # via -r requirements.in
pluggy==1.6.0
    # via
    #   pytest
    #   pytest-cov
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
pycares==5.0.1
    # via aiodns
pycparser==3.0
    # via cffi
pydantic==2.12.5
    # via
    #   fastapi
    #   sqlmodel
pydantic-core==2.41.5
    # via pydantic
pygments==2.19.2
    # via
    #   pytest
    #   rich
pytest==9.0.2
    # via
    #   -r requirements.in
    #   pytest-asyncio
    #   pytest-cov
    #   pytest-timeout
pytest-asyncio==1.3.0
    # via -r requirements.in
pytest-cov==7.0.0
    # via -r requirements.in
pytest-timeout==2.4.0
    # via kaleido
python-dateutil==2.9.0.post0
    # via
    #   croniter
    #   dateparser
    #   feedgen
    #   pandas
    #   pendulum
python-multipart==0.0.22
    # via -r requirements.in
pytz==2025.2
    # via
    #   croniter
    #   dateparser
pyyaml==6.0.3
    # via
    #   huggingface-hub
    #   transformers
redis==4.6.0
    # via
    #   fastapi-cache2
    #   rq
regex==2026.1.15
    # via
    #   dateparser
    #   transformers
rich==14.3.1
    # via
    #   -r requirements.in
    #   typer
rq==2.6.1
    # via -r requirements.in
safetensors==0.7.0
    # via transformers
scikit-learn==1.8.0
    # via -r requirements.in
scipy==1.17.0
    # via
    #   -r requirements.in
    #   scikit-learn
sgmllib3k==1.0.0
    # via feedparser
shellingham==1.5.4
    # via
    #   huggingface-hub
    #   typer
simplejson==3.20.2
    # via choreographer
six==1.17.0
    # via python-dateutil
sqlalchemy==2.0.46
    # via
    #   alembic
    #   sqlmodel
sqlmodel==0.0.31
    # via -r requirements.in
starlette==0.50.0
    # via fastapi
sympy==1.14.0
    # via torch
tenacity==9.1.2
    # via -r requirements.in
threadpoolctl==3.6.0
    # via scikit-learn
tokenizers==0.22.2
    # via transformers
torch==2.10.0
    # via -r requirements.in
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   transformers
transformers==5.0.0
    # via -r requirements.in
triton==3.6.0
    # via torch
typer==0.21.1
    # via -r requirements.in
typer-slim==0.21.1
    # via
    #   huggingface-hub
    #   transformers
typing-extensions==4.15.0
    # via
    #   alembic
    #   fastapi
    #   fastapi-cache2
    #   huggingface-hub
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   torch
    #   typer
    #   typer-slim
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.3
    # via pendulum
tzlocal==5.3.1
    # via dateparser
uvicorn==0.40.0
    # via
    #   -r requirements.in
    #   fastapi-cache2
yarl==1.22.0
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
